clean up/refactoring
