+++ common/styles/shared-styles-lit.ts
@@ -312,3 +312,11 @@ export const sharedStylesContent = `
 
 `;
 export const sharedStyles = html`${unsafeCSS(sharedStylesContent)}`;
+
+export const sharedStylesPolymer = () => {
+  const template = document.createElement('template');
+  template.innerHTML = `<style>
+    ${sharedStylesContent}
+   </style>`;
+  return template;
+};
